move watchlist functions into cryptoapi
